other_project/EAS/requirements.txt
Django==3.0.5
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-shortuuidfield==0.1.3
PyMySQL==0.9.3
requests==2.23.0
shortuuid==1.0.1
sqlparse==0.3.1
urllib3==1.25.9
chinesecalendar